KJ2007 model updated
The model now uses pre-made distributions. The model is too slow to use
in real-time (though the code to do so is retained). It now uses the
p/pdot values and chooses a pulse width from the model with closest
matching p/pdot. The .npy files are numpy arrays containing the values
for each of these models.